Modify Fast-JX initialization to work in CESM
FJX initialization involves the read-in of a some ASCII files and
a NetCDF file. Although the locations of these files are currently
hard-coded, the modifications (now being tested) should ensure
that the ASCII files are read only by the root task, with data
then broadcast to other processes. Meanwhile the NetCDF read-in
is done using PIO, which should auto-broadcast.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>